Refactor repository to seperate files
